Move ModelStub to model package in test
